+++ BankAccount.py
@@ -1,0 +1,8 @@
+class BankAccount:
+    def __init__(self, name, amount):
+        self.name = name
+        self.amount = amount
+
+    @staticmethod
+    def func(a):
+        print(a)
