--- MoneyTracker.py
@@ -1,16 +1,0 @@
-class Bank_accounts:
-    def __init__(self, name, amount):
-        self.name = name
-        self.amount = amount
-
-
-class Expenses:
-    def __init__(self, expence, cathegory) :
-        self.expense = expence
-        self.cathegory=cathegory
-
-
-class Incomes:
-    def __init__(self, income, cathegory) :
-        self.income = income
-        self.cathegory=cathegory
