+++ RunApp.py
@@ -1,0 +1,31 @@
+import BankAccount
+
+
+'''
+class BankAccounts:
+    def __init__(self, name, amount):
+        self.name = name
+        self.amount = amount
+
+
+class Expenses:
+    def __init__(self, expence, cathegory):
+        self.expense = expence
+        self.cathegory = cathegory
+
+
+class Incomes:
+    def __init__(self, income, cathegory):
+        self.income = income
+        self.cathegory = cathegory
+BankAccount.BankAccount.func(12)
+'''
+run = True
+
+while run:
+    print("Dodaj przychod - 1\n"
+          "Dodaj wydatek - 2\n"
+          "Wyswietl stan konta - 3\n"
+          "Dodaj konto - 4\n"
+          "Zakoncz - 5\n")
+    run = False
